finish did:detailed publish&update

+++ UnionSquare.Android/app/src/main/java/org/elastos/wallet/ela/ui/Assets/fragment/WalletManageFragment.java
@@ -11,19 +11,34 @@ import android.widget.TextView;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 
+import org.elastos.did.DID;
+import org.elastos.did.DIDDocument;
+import org.elastos.did.DIDStore;
+import org.elastos.did.DIDURL;
+import org.elastos.did.exception.DIDBackendException;
+import org.elastos.did.exception.DIDException;
+import org.elastos.did.exception.DIDStoreException;
 import org.elastos.wallet.R;
+import org.elastos.wallet.ela.ElaWallet.MyWallet;
 import org.elastos.wallet.ela.base.BaseFragment;
 import org.elastos.wallet.ela.bean.BusEvent;
 import org.elastos.wallet.ela.db.RealmUtil;
+import org.elastos.wallet.ela.db.table.SubWallet;
 import org.elastos.wallet.ela.db.table.Wallet;
 import org.elastos.wallet.ela.rxjavahelp.BaseEntity;
 import org.elastos.wallet.ela.rxjavahelp.NewBaseViewData;
 import org.elastos.wallet.ela.ui.Assets.fragment.mulsignwallet.ShowMulWallletPublicKeyFragment;
-import org.elastos.wallet.ela.ui.Assets.presenter.WallletManagePresenter;
+import org.elastos.wallet.ela.ui.Assets.presenter.WalletManagePresenter;
+import org.elastos.wallet.ela.ui.Assets.presenter.mulwallet.CreatMulWalletPresenter;
 import org.elastos.wallet.ela.ui.Assets.viewdata.WalletManageViewData;
 import org.elastos.wallet.ela.ui.common.bean.CommmonBooleanEntity;
+import org.elastos.wallet.ela.ui.common.bean.CommmonObjEntity;
 import org.elastos.wallet.ela.ui.common.bean.CommmonStringEntity;
+import org.elastos.wallet.ela.ui.common.bean.ISubWalletListEntity;
 import org.elastos.wallet.ela.ui.common.viewdata.CommmonStringWithMethNameViewData;
+import org.elastos.wallet.ela.ui.did.fragment.AddDIDFragment;
+import org.elastos.wallet.ela.ui.did.fragment.DidDetailFragment;
+import org.elastos.wallet.ela.ui.did.presenter.AddDIDPresenter;
 import org.elastos.wallet.ela.utils.DialogUtil;
 import org.elastos.wallet.ela.utils.RxEnum;
 import org.elastos.wallet.ela.utils.listener.WarmPromptListener;
@@ -31,17 +46,19 @@ import org.elastos.wallet.ela.utils.listener.WarmPromptListener2;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import butterknife.BindView;
 import butterknife.OnClick;
 import butterknife.Unbinder;
 
-public class WallletManageFragment extends BaseFragment implements WarmPromptListener, WalletManageViewData, CommmonStringWithMethNameViewData, NewBaseViewData {
+public class WalletManageFragment extends BaseFragment implements WarmPromptListener, WalletManageViewData, CommmonStringWithMethNameViewData, NewBaseViewData {
 
     private static final String DELETE = "delete";
 
     private static final String OUTPORTMN = "outportmm";
+    private static final String DIDINIT = "didinit";
     private static final String OUTPORTMUPK = "outportmupk";
     @BindView(R.id.tv_title)
     TextView tvTitle;
@@ -58,6 +75,8 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     LinearLayout llExportmnemonic;
     @BindView(R.id.ll_sign)
     LinearLayout llSign;
+    @BindView(R.id.ll_did)
+    LinearLayout llDid;
     @BindView(R.id.ll_exportreadonly)
     LinearLayout llExportreadonly;
     @BindView(R.id.ll_showmulpublickey)
@@ -68,8 +87,9 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     String dialogAction = null;
     private Dialog dialog;
     private Wallet wallet;
-    private WallletManagePresenter presenter;
+    private WalletManagePresenter presenter;
     private String payPasswd;
+    private DIDStore store;
 
     @Override
     protected int getLayoutId() {
@@ -90,7 +110,7 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     protected void initView(View view) {
         tvTitle.setText(R.string.walletmanage);
         dialogUtil = new DialogUtil();
-        presenter = new WallletManagePresenter();
+        presenter = new WalletManagePresenter();
         registReceiver();
         switch (wallet.getType()) {
             //0 普通单签 1单签只读 2普通多签 3多签只读
@@ -98,16 +118,19 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                 llShowwalletpublickey.setVisibility(View.GONE);
                 break;
             case 1:
+                llDid.setVisibility(View.GONE);
                 llUpdatepwd.setVisibility(View.GONE);
                 llExportmnemonic.setVisibility(View.GONE);
                 llShowwalletpublickey.setVisibility(View.GONE);
                 break;
             case 2:
+                llDid.setVisibility(View.GONE);
                 llExportmnemonic.setVisibility(View.GONE);
                 llExportreadonly.setVisibility(View.GONE);
                 llShowmulpublickey.setVisibility(View.GONE);
                 break;
             case 3:
+                llDid.setVisibility(View.GONE);
                 llUpdatepwd.setVisibility(View.GONE);
                 llExportmnemonic.setVisibility(View.GONE);
                 llExportreadonly.setVisibility(View.GONE);
@@ -117,7 +140,7 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     }
 
     @OnClick({R.id.tv_delete, R.id.ll_updatename, R.id.ll_updatepwd, R.id.ll_exportkeystore, R.id.ll_exportmnemonic,
-            R.id.ll_sign, R.id.ll_exportreadonly, R.id.ll_showmulpublickey, R.id.ll_showwalletpublickey, R.id.ll_nodeconect})
+            R.id.ll_sign, R.id.ll_exportreadonly, R.id.ll_showmulpublickey, R.id.ll_showwalletpublickey, R.id.ll_nodeconect, R.id.ll_did})
     public void onViewClicked(View view) {
         Bundle bundle = null;
         dialogAction = null;
@@ -127,7 +150,7 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                 dialogUtil.showWarmPrompt1(getBaseActivity(), getString(R.string.deletewallletornot), new WarmPromptListener() {
                     @Override
                     public void affireBtnClick(View view) {
-                        dialog = dialogUtil.showWarmPromptInput(getBaseActivity(), null, null, WallletManageFragment.this);
+                        dialog = dialogUtil.showWarmPromptInput(getBaseActivity(), null, null, WalletManageFragment.this);
                         dialogAction = DELETE;
                     }
                 });
@@ -180,9 +203,12 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                 break;
             case R.id.ll_nodeconect:
                 //节点连接设置
-
-
                 start(NodeConnectSetFragment.class, getArguments());
+                break;
+            case R.id.ll_did:
+                //did
+                new AddDIDPresenter().getAllSubWallets(wallet.getWalletId(), this);
+
                 break;
 
         }
@@ -191,29 +217,36 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
 
     @Override
     public void affireBtnClick(View view) {
-//这里只见他showWarmPromptInput的确认
-        if (wallet.getType() == 1 || wallet.getType() == 3) {
+        //这里只见他showWarmPromptInput的确认
+        if (DELETE.equals(dialogAction) && (wallet.getType() == 1 || wallet.getType() == 3)) {
             //0 普通单签 1单签只读 2普通多签 3多签只读
+            //不需要密码
             presenter.destroyWallet(wallet.getWalletId(), this);
             return;
         }
+        //下面都需要密码
         payPasswd = ((EditText) view).getText().toString().trim();
         if (TextUtils.isEmpty(payPasswd)) {
             showToastMessage(getString(R.string.pwdnoempty));
             return;
         }
+        switch (dialogAction) {
+            case DELETE:
+            case OUTPORTMUPK:
+            case DIDINIT:
+                //删除钱包  验证密码
+                //查看多签公钥的兼容  验证密码
+                //初始化did
+                presenter.verifyPayPassword(wallet.getWalletId(), payPasswd, this);
+                break;
 
-        if (OUTPORTMN.equals(dialogAction)) {
-            //导出助记词
-            presenter.exportWalletWithMnemonic(wallet.getWalletId(), payPasswd, this);
 
-        } else if (DELETE.equals(dialogAction)) {
-            //删除钱包  验证密码
-            presenter.verifyPayPassword(wallet.getWalletId(), payPasswd, this);
-        } else if (OUTPORTMUPK.equals(dialogAction)) {
-            //查看多签公钥的兼容  输一遍密码
-            presenter.verifyPayPassword(wallet.getWalletId(), payPasswd, this);
+            case OUTPORTMN:
+                //导出助记词
+                presenter.exportWalletWithMnemonic(wallet.getWalletId(), payPasswd, this);
+                break;
         }
+
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
@@ -224,8 +257,40 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
             tvUpdatename.setText(result.getName());
 
         }
+        if (integer == RxEnum.ADDPROPERTY.ordinal()) {
+            //增加子钱包后
+            //did初始化获得密码
+            dialog = dialogUtil.showWarmPromptInput(getBaseActivity(), null, null, this);
+            dialogAction = DIDINIT;
+        }
     }
 
+    private void initDid() {
+
+        try {
+            store = getMyDID().getDidStore();
+            if (store.containsPrivateIdentity()) {
+                resolve();
+            } else {
+                //获得私钥用于初始化did
+                new CreatMulWalletPresenter().exportxPrivateKey(wallet.getWalletId(), payPasswd, this);
+            }
+        } catch (DIDException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    private void resolve() throws DIDStoreException, DIDBackendException {
+        DID did = getMyDID().initDID(payPasswd);
+        DIDURL didurl = new DIDURL(did, "primary");
+
+        if (!store.containsDid(did) || !store.containsPrivateKey(did, didurl)) {
+            store.newDid(0, payPasswd);
+        }
+        presenter.DIDResolve(did, this, null);
+
+    }
 
     @Override
     public void onDestoryWallet(String data) {
@@ -247,11 +312,26 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     @Override
     public void onGetCommonData(String methodname, String data) {
         //exportWalletWithMnemonic
-        dialog.dismiss();
-        Bundle bundle = new Bundle();
-        bundle.putString("mnemonic", data);
-        bundle.putParcelable("wallet", wallet);
-        start(OutportMnemonicFragment.class, bundle);
+        switch (methodname) {
+
+            case "exportWalletWithMnemonic":
+                dialog.dismiss();
+                Bundle bundle = new Bundle();
+                bundle.putString("mnemonic", data);
+                bundle.putParcelable("wallet", wallet);
+                start(OutportMnemonicFragment.class, bundle);
+                break;
+            case "exportxPrivateKey":
+                dialog.dismiss();
+                try {
+                    store.initPrivateIdentity(data, payPasswd);
+                    resolve();
+                } catch (DIDException e) {
+                    e.printStackTrace();
+                    showToast(getString(R.string.didinitfaile));
+                }
+                break;
+        }
 
     }
 
@@ -259,6 +339,38 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
     @Override
     public void onGetData(String methodName, BaseEntity baseEntity, Object o) {
         switch (methodName) {
+            case "DIDResolve":
+                Bundle bundle1 = new Bundle();
+                bundle1.putParcelable("wallet", wallet);
+                DIDDocument doc = (DIDDocument) ((CommmonObjEntity) baseEntity).getData();
+                if (doc == null) {
+                    //doc 为空  创建did
+                    start(AddDIDFragment.class, bundle1);
+                } else {
+                    //展示did详情
+                    try {
+                        store.storeDid(doc);//存储本地
+                        start(DidDetailFragment.class, bundle1);
+                    } catch (DIDStoreException e) {
+                        e.printStackTrace();
+                    }
+
+                }
+                break;
+            case "getAllSubWallets":
+                ISubWalletListEntity subWalletListEntity = (ISubWalletListEntity) baseEntity;
+                for (SubWallet subWallet : subWalletListEntity.getData()) {
+                    if (subWallet.getChainId().equals(MyWallet.IDChain)) {
+                        //did初始化获得密码
+                        dialog = dialogUtil.showWarmPromptInput(getBaseActivity(), null, null, this);
+                        dialogAction = DIDINIT;
+                        return;
+                    }
+                }
+                //没有对应的子钱包 需要打开idchain
+                showOpenDIDWarm(subWalletListEntity);
+
+                break;
             case "getPubKeyInfo":
                 String pubKeyInfo = ((CommmonStringEntity) baseEntity).getData();
                 JSONObject pubKeyInfoJsonData = JSON.parseObject(pubKeyInfo);
@@ -290,14 +402,18 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                 }
                 break;
             case "verifyPayPassword":
-                //目前只在删除一种情况调用
+
                 boolean result = ((CommmonBooleanEntity) baseEntity).getData();
                 if (result) {
                     dialog.dismiss();
                     if (OUTPORTMUPK.equals(dialogAction)) {
+                        //查看多签公钥
                         presenter.getPubKeyInfo(wallet.getWalletId(), this);
                     } else if (DELETE.equals(dialogAction)) {
+                        //目前只在删除一种情况调用
                         presenter.getMasterWalletBasicInfo(wallet.getWalletId(), this);
+                    } else if (DIDINIT.equals(dialogAction)) {
+                        initDid();
                     }
                 } else {
                     showToastMessage(getString(R.string.error_20003));
@@ -322,13 +438,13 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                                 showToastMessage(getString(R.string.pwdnoempty));
                                 return;
                             }
-                            presenter.verifyPassPhrase(wallet.getWalletId(), passphrase, payPasswd, WallletManageFragment.this);
+                            presenter.verifyPassPhrase(wallet.getWalletId(), passphrase, payPasswd, WalletManageFragment.this);
                         }
 
                         @Override
                         public void noAffireBtnClick(View view) {
                             //直接删除
-                            presenter.destroyWallet(wallet.getWalletId(), WallletManageFragment.this);
+                            presenter.destroyWallet(wallet.getWalletId(), WalletManageFragment.this);
                         }
                     });
 
@@ -343,5 +459,23 @@ public class WallletManageFragment extends BaseFragment implements WarmPromptLis
                 break;
         }
 
+    }
+
+    private void showOpenDIDWarm(ISubWalletListEntity subWalletListEntity) {
+        new DialogUtil().showCommonWarmPrompt(getBaseActivity(), getString(R.string.noidchainopenornot),
+                getString(R.string.toopen), getString(R.string.cancel), false, new WarmPromptListener() {
+                    @Override
+                    public void affireBtnClick(View view) {
+                        Bundle bundle = new Bundle();
+
+                        ArrayList<String> chainIds = new ArrayList<>();
+                        for (SubWallet iSubWallet : subWalletListEntity.getData()) {
+                            chainIds.add(iSubWallet.getChainId());
+                            bundle.putString("walletId", iSubWallet.getBelongId());
+                        }
+                        bundle.putStringArrayList("chainIds", chainIds);
+                        start(AddAssetFragment.class, bundle);
+                    }
+                });
     }
 }
